move database to typescript
